27. Creating & Handling Events
Creating the Event

Commands, events & their handlers look identical. In the src/Message directory, to start organizing things a bit better, let's create an Event/ subdirectory. Inside, add a new class: ImagePostDeletedEvent.

Notice the name of this class: that's critical. Everything so far has sounded like a command: we're running around our code base shouting: AddPonkaToImage! And DeleteImagePost! We sound bossy.

But with events, you're not using a strict command, you're notifying the system of something that just happened: we're going to fully delete the image post and then say:

 Hey! I just deleted an image post! If you care... uh... now is your chance to... uh... do something! But I don't care if you do or not.

The event itself could be handled by... nobody... or it could have multiple handlers. Inside the class, we'll store any data we think might be handy. Add a constructor with a string $filename - knowing the filename of the deleted ImagePost might be useful. I'll hit Alt + Enter and go to "Initialize Fields" to create that property and set it. Then, at the bottom, I'll go to "Code -> Generate" - or Command + N on a Mac - and select "Getters" to generate this one getter.

You may have noticed that, other than its name, this "event" class looks exactly like the command we just deleted!

Creating the Event Handler

Creating an event "handler" also looks identical to command handlers. In the MessageHandler directory, let's create another subdirectory called Event/ for organization. Then add a new PHP class. Let's call this RemoveFileWhenImagePostDeleted. Oh... but make sure you spell that all correctly.

This also follows a different naming convention. For commands, if a command was named AddPonkaToImage, we called the handler AddPonkaToImageHandler. The big difference between commands and events is that, while each command has exactly one handler - so using the "command name Handler" convention makes sense - each event could have multiple handlers.

But the inside of a handler looks the same: implement MessageHandlerInterface and then create our beloved public function __invoke() with the type-hint for the event class: ImagePostDeletedEvent $event.

Now... we'll do the work... and this will be identical to the handler we just deleted. Add a constructor with the one service we need to delete files: PhotoFileManager. I'll initialize fields to create that property then, down below, finish things with $this->photoFileManager->deleteImage() passing that $event->getFilename().

I hope this was delightfully boring for you. We deleted a command and command handler... and replaced them with an event and an event handler that are... other than the name... identical!

Next, let's dispatch this new event... but to our event bus. Then, we'll tweak that bus a little bit to make sure it works perfectly.