32.Query Bus The last type of bus that you'll hear about is... the double-decker tourist bus! I mean... the query bus! Full disclosure... while I am a fan of waving like an idiot on the top-level of a tourist bus, I'm not a huge fan of query buses: I think they make your code a bit more complex... for not much benefit. That being said, I want you to at least understand what it is and how it fits into the message bus methodology.
Creating the Query Bus

In config/packages/messenger.yaml we have command.bus and event.bus. Let's add query.bus. I'll keep things simple and just set this to ~ to get the default settings.

What is a Query?

Ok: so what is the point of a "query bus"? We understand the purpose of commands: we dispatch messages that sound like commands: AddPonkaToImage or DeleteImagePost. Each command then has exactly one handler that performs that work... but doesn't return anything. I haven't really mentioned that yet: commands just do work, but they don't communicate anything back. Because of this, it's ok to process commands synchronously or asynchronously - our code isn't waiting to get information back from the handler.

A query bus is the opposite. Instead of commanding the bus to do work, the point of a query is to get information back from the handler. For example, let's pretend that, on our homepage, we want to print the number of photos that have been uploaded. This is a question or query that we want to ask our system:

  How many photos are in the database?

If you're using the query bus pattern, instead of getting that info directly, you'll dispatch a query.

Creating the Query & Handler

Inside the Message/ directory, create a new Query/ subdirectory. And inside of that, create a new PHP class called GetTotalImageCount.

Even that name sounds like a query instead of a command: I want to get the total image count. And... in this case, we can leave the query class blank: we won't need to pass any extra data to the handler.

Next, inside of MessageHandler/, do the same thing: add a Query/ subdirectory and then a new class called GetTotalImageCountHandler. And like with everything else, make this implement MessageHandlerInterface and create public function __invoke() with an argument type-hinted with the message class: GetTotalImageCount $getTotalImageCount.

What do we do inside of here? Find the image count! Probably by injecting the ImagePostRepository, executing a query and then returning that value. I'll leave the querying part to you and just return 50.

But hold on a second... cause we just did something totally new! We're returning a value from our handler! This is not something that we've done anywhere else. Commands do work but don't return any value. A query doesn't really do any work, its only point is to return a value.

Before we dispatch the query, open up config/packages/services.yaml so we can do our same trick of binding each handler to the correct bus. Copy the Event\ section, paste, change Event to Query in both places... then set the bus to query.bus.

Love it! Let's check our work by running:

php bin/console debug:messenger
Yep! query.bus has one handler, event.bus has one handler and command.bus has two.