32.Query Bus Dispatching the Message
Open up src/Controller/MainController.php. This renders the homepage and so this is where we need to know how many photos have been uploaded. To get the query bus, we need to know which type-hint & argument name combination to use. We get that info from running:

php bin/console debug:autowiring mess
We can get the main command.bus by using the MessageBusInterface type-hint with any argument name. To get the query bus, we need to use that type-hint and name the argument: $queryBus.

Do that: MessageBusInterface $queryBus. Inside the function, say $envelope = $queryBus->dispatch(new GetTotalImageCount()).

Fetching the Returned Value

We haven't used it too much, but the dispatch() method returns the final Envelope object, which will have a number of different stamps on it. One of the properties of a query bus is that every query will always be handled synchronously. Why? Simple: we need the answer to our query... right now! And so, our handler must be run immediately. In Messenger, there's nothing that enforces this on a query bus... it's just that we won't ever route our queries to a transport, so they'll always be handled right now.

Anyways, once a message is handled, Messenger automatically adds a stamp called HandledStamp. Let's get that: $handled = $envelope->last() with HandledStamp::class. I'll add some inline documentation above that to tell my editor that this will be a HandledStamp instance.

So... why did we get this stamp? Well, we need to know what the return value of our handler was. And, conveniently, Messenger stores that on this stamp! Get it with $imageCount = $handled->getResult().

Let's pass that into the template as an imageCount variable.... and then in the template - templates/main/homepage.html.twig - because our entire frontend is built in Vue.js, let's override the title block on the page and use it there: Ponka'd {{ imageCount }} Photos.

Let's check it out! Move over, refresh and... it works! We've Ponka's 50 photos... at least according to our hardcoded logic.

So... that's a query bus! It's not my favorite because we're not guaranteed what type it returns - the imageCount could really be a string... or an object of any class. Because we're not calling a direct method, the data we get back feels a little fuzzy. Plus, because queries need to be handled synchronously, you're not saving any performance by leveraging a query bus: it's purely a programming pattern.

But, my opinion is totally subjective, and a lot of people love query buses. In fact, we've been talking mostly about the tools themselves: command, event & query buses. But there are some deeper patterns like CQRS or event sourcing that these tools can unlock. This is not something we currently use here on SymfonyCasts... but if you're interested, you can read more about this topic - Matthias Noback's blog is my favorite source.

Oh, and before I forget, if you look back on the Symfony docs... back on the main messenger page... all the way at the bottom... there's a spot here about getting results from your handler. It shows some shortcuts that you can use to more easily get the value from the bus.

Next, let's talk about message handler subscribers: an alternate way to configure a message handler that has a few extra options.