35. Advanced Handler Config: Handler Subscribers Controlling the Method & Handling Multiple Classes
Ok... but since we probably should use type-hints... this isn't that interesting yet. What else can we do?

Well, by assigning this to an array, we can add some config. For example, we can say 'method' => '__invoke'. Yep, we can now control which method Messenger will call. That's especially useful if you decide that you want to add another yield to handle a second message... and want Messenger to call a different method.

Handler Priority

What else can we put here? One option is priority - let's set it to... 10. This option is... much less interesting than it might look like at first. We talked earlier about priority transports: in config/packages/messenger.yaml we created two transport - async & async_priority_high - and we route different messages to each. We did this so that, when we run our worker, we can tell it to always read messages from async_priority_high first before reading messages from async. That makes async_priority_high a place for us to send "higher" priority messages.

The priority option here is... less powerful. If you send a message to a transport with a priority 0 and then you send another message to that same transport with priority 10, what do you think will happen? Which message will be handled first?

The answer: the first message that was sent - the one with the lower priority. Basically, Messenger will always read messages in a first-in-first-out basis: it will always read the oldest messages first. The priority does not influence this.

So... what does it do? Well, if DeleteImagePost had two handlers... and one had the default priority of zero and another had 10, the handler with priority 10 would be called first. That's not usually important, but could be if you had two event handlers and really needed them to happen in a certain order.

Next, let's talk about one more option you can pass here - the most powerful option. It's called from_transport and allows you to, sort of, send different "handlers" of a message to different transports so that each can be consumed independently.