34. Sending Handlers to Different Transports: from_transport The last option I want to mention is interesting... but can also be confusing. It's called from_transport.
If you look at messenger.yaml, this DeleteImagePost isn't being routed anywhere, which means it's handled synchronously. Let's pretend that we want to handle it asynchronously... and that we're routing it to the async transport. Set from_transport to async... then temporarily route this class to that transport in messenger.yaml.

Now, pretend that the DeleteImagePost message actually has two handlers... something that's very possible for events. Assuming that we did not add this from_transport config yet, if you sent DeleteImagePost to the async transport, then when that message is read from that transport by a worker, both handlers will be executed one after another.

But what if you wanted to, sort of, send one handler of that message to one transport, maybe async_priority_high, and another handler to another transport. Well, in Messenger, you don't send "handlers"... you send messages... and when Messenger consumes a message, it calls all the handlers for that message.